+++ src/components/home/LandPage.jsx
@@ -5,23 +5,17 @@ import Button from '@mui/material/Button';
 import Image from 'next/image';
 import { useState } from 'react';
 
+import AppLogo from '@/components/app/AppLogo';
+import FZDialog from '@/components/misc/Dialog';
 import Link from '@/components/misc/Link';
 import mainImage from '@/images/mainImage2.png';
-
-import FZDialog from '../misc/Dialog';
 
 const LandPage = () => {
   return (
     <Stack className='w-full xl:flex-row h-full p-2'>
       <section className='flex flex-col gap-3 px-3 h-full items-center smh:justify-center'>
         <Stack>
-          <Typography
-            variant='h1'
-            className=' m-0 p-0 flex gap-2 justify-center items-end text-xl font-bold foont-mono text-center text-white '
-          >
-            <span className='text-3xl text-amber-500'>SHIKOMORI</span>
-            <small>Taradjam</small>
-          </Typography>
+          <ResponsiveAppLogo />
           <Typography
             variant='body2'
             className='text-[0.98rem] font-mono text-center text-slate-100'
@@ -42,7 +36,7 @@ const LandPage = () => {
           </Button>
           <ExplainModal />
         </Stack>
-        <Stack gap={1} className=' md:max-w-[90%] items-center '>
+        <Stack gap={1} className='md:max-w-[90%] items-center '>
           <Typography
             variant='body2'
             className='font-serif text-center text-xl text-slate-100'
@@ -64,6 +58,22 @@ const LandPage = () => {
 };
 
 export default LandPage;
+function ResponsiveAppLogo() {
+  return (
+    <>
+      <AppLogo size='medium' sx={{ display: { xs: 'block', sm: 'none' } }} />
+      <AppLogo
+        size='xlarge'
+        sx={{ display: { xs: 'none', sm: 'block', md: 'none' } }}
+      />
+      <AppLogo
+        size='2xl'
+        sx={{ display: { xs: 'none', sm: 'none', md: 'block' } }}
+      />
+    </>
+  );
+}
+
 function ExplainModal() {
   const [open, setOpen] = useState(false);
   return (
